fixed export and cd, mv search and delete to utils in mish and create file hashmap_search_key
